pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = '^3.13'
fastapi = "^0.100.0"
uvicorn = {extras = ["standart"], version = "^0.34.0"}
pydantic = {extras = ["email"], version = "^2.10.6"}
pydantic-settings = "^2.7.0"
python-dotenv = "^1.0.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
envparse = "^0.2.0"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
passlib = "^1.7.4"
bcrypt = "^4.2.1"
python-jose = "^3.3.0"
httpx = "^0.28.1"
redis = "^5.2.1"
setuptools = "^75.8.0"
aioredis = "2.0.0"

psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.20"
pdfkit = "^1.0.0"
aiosmtplib = "^4.0.0"
reportlab = "^4.3.1"
celery = "^5.4.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
